requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
authlib==1.2.1
itsdangerous==2.1.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.0
python-dotenv==1.0.0
aiosmtplib==3.0.1
jinja2==3.1.2
python-jose[cryptography]==3.3.0
